requirements.txt
Flask
opencv-python
numpy
scikit-learn
imutils
requests
python-dotenv
Flask-SocketIO 